Remove clang problematic cast to uint32_t for pointer

+++ src/stm32/usb/usbd_conf.c
@@ -315,7 +315,7 @@ void HAL_PCD_SuspendCallback(PCD_HandleTypeDef *hpcd)
   /*Enter in STOP mode */
   if (hpcd->Init.low_power_enable) {
     /* Set SLEEPDEEP bit and SleepOnExit of Cortex System Control Register */
-    SCB->SCR |= (uint32_t)((uint32_t)(SCB_SCR_SLEEPDEEP_Msk | SCB_SCR_SLEEPONEXIT_Msk));
+    SCB->SCR |= (uint32_t)((uintptr_t)(SCB_SCR_SLEEPDEEP_Msk | SCB_SCR_SLEEPONEXIT_Msk));
   }
 }
